Hack `writable_attributes` to allow `status` as a writable property
In order to have the attribute named `status_code` in the contract
and a differently named property in the Representer and still be able
to be marked as writable, I needed to add this hack in.

I'd like to instead be able to provide an alias in the representer
property such as:

```
resource :status_code,
         as: :status
         ...
         ...
```

and a contract which defines the attribute as

```
attribute :status_code
```

which then allows `status` to be marked as writeable and be responded
with in the PayloadRepresenter.